Refactor eleventy code into folder src/eleventy
