+++ src/utils/path.js
@@ -1,21 +1,21 @@
 const portfolio = require('../data/portfolio.json')
 
-const getRootPath = (root) => process.argv.includes('--serve') ? '' : root
+const getRootPath = () => process.argv.includes('--serve') ? '' : portfolio.root
 
-const getRelativePath = (root, path, lang) => {
+const getRelativePath = (path, lang) => {
   if (!lang || lang === portfolio.i18n.default) {
-    return getRoot(root) + path
+    return getRootPath() + path
   }
 
-  return `${getRoot(root)}/${lang}${path}`
+  return `${getRootPath()}/${lang}${path}`
 }
 
-const getAssetPath = (root, assetType, asset) => {
+const getAssetPath = (assetType, asset) => {
   if (asset.search('http') !== -1) {
     return asset
   }
   
-  return `${getRoot(root)}/assets/${assetType}/${asset}`
+  return `${getRootPath()}/assets/${assetType}/${asset}`
 }
 
 module.exports = {
